+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.34.4] - 2023-06-27
+
 ### Fixed
 
 - Postal code rules when using geolocation mode